backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

motor==3.4.0
beanie==1.25.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

httpx==0.27.0

beautifulsoup4==4.12.3
lxml==5.2.1

redis==5.0.4

celery==5.4.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
respx==0.21.1            
ruff==0.4.4